refactor:  create custum hooks usemenu
