+++ src/components/pages/order/OrderPage.jsx
@@ -6,15 +6,13 @@ import Main from "./MainMenu/Main";
 import theme from "../../../theme/index";
 import OrderContext from "../../../context/OrderContext";
 import EMPTY_PRODUCT from "../../../enums/product";
-import { deepClone } from "../../../utils/array";
-import { fakeMenu } from "../../../fakeData/FakeMenu";
-import { toast } from "react-toastify";
+
+import UseMenu from "../../../hooks/UseMenu";
 
 
 function OrderPage() {
   const [isAdminMode, setAdminMode] = useState(false);
   const [isCollapsed, setIsCollapsed] = useState(false);
-  const [menuData, setMenuData] = useState(fakeMenu.LARGE);
   const [productSelected, setProductSelected] = useState({ EMPTY_PRODUCT });
   const titleEditRef = useRef()
 
@@ -22,41 +20,7 @@ function OrderPage() {
 
   const navigate = useNavigate();
 
-  const addProductToMenu = (newProduct) => {
-    setMenuData([...menuData, newProduct]);
-  };
-
-  const handleEdit = (productBeingEdited) => {
-    const indexOfProductToEdit = menuData.findIndex(
-      (product) => product.id === productBeingEdited.id
-    );
-  
-    if (indexOfProductToEdit !== -1) {
-      const updatedMenu = deepClone(menuData);
-      updatedMenu[indexOfProductToEdit] = productBeingEdited;
-      setMenuData(updatedMenu);
-    }
-  };
-
-  const resetMenu = () => {
-    setMenuData(fakeMenu.LARGE)
-    toast.success('Produits regénéré avec success')
-
-  }
-
-  const handleDelete = (idOfProductToDelete) => {
-    //1. copy du state
-    const menuCopy = deepClone(menuData)
-
-    //2. manip de la copie state
-    const menuUpdated = menuCopy.filter((product) => product.id !== idOfProductToDelete)
-    console.log("menuUpdated: ", menuUpdated)
-
-    //3. update du state
-    setMenuData(menuUpdated)
-    toast.success('Supprimé avec success')
-  }
-  
+  const {handleEdit, handleDelete,resetMenu,addProductToMenu,menuData,setMenuData} = UseMenu()
   const orderContextValue = {
     isAdminMode,
     setAdminMode,
